python/requirements.txt
# EyeLab Python Dependencies — Phase 1 Webcam MVP
# Install: pip install -r requirements.txt
# Python 3.9+ required

# Core scientific libraries
numpy>=1.21.0
scipy>=1.7.0

# Computer vision — ArUco moved to objdetect in 4.8+; 4.11+ recommended
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0  # includes aruco module

# Modal analysis — UNV/UFF file I/O
pyuff>=1.0.0

# Data processing and plotting
pandas>=1.3.0
matplotlib>=3.4.0

# Configuration file support
pyyaml>=5.4.0

# GUI — image display in tkinter
Pillow>=9.0.0
